Change units on a graph?
I opened a model during a class demo, and I don't recall changing anything about it ... then it seems like it changed the axes of the plot from acre-ft/day to cms. I wonder if there's an action somewhere in RiverWare that preserves the default units, or overwrites the unit scheme, and that gets triggered during a save?