Improve plot in two-res model; initial commit of seasonal model
The seasonal model was successfully created where the rule curve zones are looked up in a monthly table. Right now the model isn't interesting because the time horizon is not very long and it doesn't span when the seasonal change in the table does. But at least the logic seems to work!