DMIs to import USGS data in Series-Parallel example, wrote initial rules
The series and parallel example is running with very basic functionality. The reservoirs aren't doing much, but everything is working and tested for basic operations -- releasing constant flowrate, handling dead storage, and flood releases (although not explicitly tested). This will be a good stepping point to creating example models for class.